+++ demo/src/test/java/Playground.java
@@ -1,9 +1,13 @@
 import io.github.palexdev.materialfx.controls.MFXButton;
 import io.github.palexdev.materialfx.controls.MFXTextField;
+import io.github.palexdev.materialfx.controls.MFXToggleButton;
+import io.github.palexdev.materialfx.font.MFXFontIcon;
 import javafx.application.Application;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
+import javafx.scene.control.ContentDisplay;
 import javafx.scene.layout.VBox;
+import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 import org.scenicview.ScenicView;
 
@@ -25,7 +29,11 @@ public class Playground extends Application {
 			textField.setMeasureUnit(measureUnit);
 		});
 
-		vBox.getChildren().addAll(button, textField);
+		MFXToggleButton tb = new MFXToggleButton("Text", MFXFontIcon.getRandomIcon(24, Color.BLACK));
+		tb.setContentDisposition(ContentDisplay.GRAPHIC_ONLY);
+		tb.setContentDisplay(ContentDisplay.GRAPHIC_ONLY);
+
+		vBox.getChildren().addAll(button, textField, tb);
 		Scene scene = new Scene(vBox, 800, 800);
 		primaryStage.setScene(scene);
 		primaryStage.show();